Moved functions to dedicated service class
