+++ curation/scripts/replace_obsolete_traits.py
@@ -3,6 +3,7 @@ import sys
 
 from catalog.models import EFOTrait, EFOTrait_Ontology
 from core.services.ols_rest_client import OLSRestClient
+import curation.services.efo_trait_service as efo_trait_service
 from auditlog.context import set_extra_data
 
 # Note: This script is intended to be run via Django's manage.py shell, e.g.:
@@ -13,37 +14,13 @@ from auditlog.context import set_extra_data
 EXPORTED_SCORE_IDS_FILENAME = 'updated_score_ids.txt'
 
 
-def get_or_make_efo_trait(trait_efo_id) -> EFOTrait:
-    """Fetches an EFOTrait by its EFO ID, or creates it if it doesn't exist in the database and fetches its info
-    from the OLS REST API."""
-    efo_trait, created = EFOTrait.objects.get_or_create(id=trait_efo_id)
-    if created:
-        efo_trait.parse_api()
-        efo_trait.save()
-        print(f'Created new EFOTrait "{efo_trait.label}" ({efo_trait.id}).')
-    return efo_trait
-
-
-def clear_ontology_tables():
-    """Clears all entries from the EFOTrait_Ontology tables, so they can be regenerated by the UpdateEFO script
-    during release."""
-    EFOTrait_Ontology.objects.all().delete()
-    print('Cleared all EFOTrait_Ontology entries from the database.')
-
-
 def replace_score_trait_in_scores(obsolete_efo_trait: EFOTrait, replacement_efo_trait: EFOTrait) -> set[str]:
     """Replaces the obsolete EFO trait with the replacement EFO trait in all associated scores.
     Returns a set of updated score IDs for reference."""
     print(f'Replacing obsolete trait {obsolete_efo_trait.id} with {replacement_efo_trait.id} in:')
-    print('  - ', end='')
-    updated_score_ids = set()
     with set_extra_data({'additional_data': {'reason': 'Obsolete trait replacement'}}):  # extra auditlog info
-        for score in obsolete_efo_trait.associated_scores.all():
-            print(f'{score.id}', end=', ')
-            score.trait_efo.remove(obsolete_efo_trait)
-            score.trait_efo.add(replacement_efo_trait)
-            updated_score_ids.add(score.id)
-    print('')
+        updated_score_ids = efo_trait_service.replace_score_trait_in_scores(obsolete_efo_trait, replacement_efo_trait)
+    print(f'  - {len(updated_score_ids)} scores: {", ".join(sorted(updated_score_ids))}')
     return updated_score_ids
 
 
@@ -62,7 +39,7 @@ def main():
                 print(f'ALERT: No replacement term specified for obsolete trait {trait_efo.id}. Aborting.')
                 return 1
             replacement_term_id = replacement_term_id.split('/')[-1]
-            replacement_efo_trait = get_or_make_efo_trait(replacement_term_id)
+            replacement_efo_trait = efo_trait_service.get_or_make_efo_trait(replacement_term_id)
             updated_score_ids.update(replace_score_trait_in_scores(trait_efo, replacement_efo_trait))
 
     print(f'Finished processing. {len(obsolete_traits)} obsolete traits were found and updated with their replacements where applicable.')
@@ -71,7 +48,7 @@ def main():
     with open(EXPORTED_SCORE_IDS_FILENAME, 'w') as f:
         for score_id in sorted(updated_score_ids):
             f.write(f'{score_id}\n')
-    print(f'Exported {len(updated_score_ids)} updated score IDs to "{EXPORTED_SCORE_IDS_FILENAME}".')
+    print(f'>> Exported {len(updated_score_ids)} updated score IDs to "{EXPORTED_SCORE_IDS_FILENAME}".')
 
     return 0
