+++ curation/services/efo_trait_service.py
@@ -1,0 +1,30 @@
+from catalog.models import EFOTrait, EFOTrait_Ontology
+
+
+def get_or_make_efo_trait(trait_efo_id) -> EFOTrait:
+    """Fetches an EFOTrait by its EFO ID, or creates it if it doesn't exist in the database and fetches its info
+    from the OLS REST API."""
+    efo_trait, created = EFOTrait.objects.get_or_create(id=trait_efo_id)
+    if created:
+        efo_trait.parse_api()
+        efo_trait.save()
+        print(f'Created new EFOTrait "{efo_trait.label}" ({efo_trait.id}).')
+    return efo_trait
+
+
+def replace_score_trait_in_scores(obsolete_efo_trait: EFOTrait, replacement_efo_trait: EFOTrait) -> set[str]:
+    """Replaces the obsolete EFO trait with the replacement EFO trait in all associated scores.
+    Returns a set of updated score IDs for reference."""
+    updated_score_ids = set()
+    for score in obsolete_efo_trait.associated_scores.all():
+        score.trait_efo.remove(obsolete_efo_trait)
+        score.trait_efo.add(replacement_efo_trait)
+        updated_score_ids.add(score.id)
+    return updated_score_ids
+
+
+def clear_ontology_tables():
+    """Clears all entries from the EFOTrait_Ontology tables, so they can be regenerated by the UpdateEFO script
+    during release."""
+    EFOTrait_Ontology.objects.all().delete()
+    print('Cleared all EFOTrait_Ontology entries from the database.')
